requirements.txt
google-generativeai>=0.8.0
Pillow
python-dotenv
markdown2
weasyprint
